fastapi_worker/requirements.txt
aio-pika>=9.4.0
beautifulsoup4>=4.12.3
fastapi>=0.115.0
langfuse<=3.0.0
openai>=1.85.0,<1.92.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
requests>=2.32.0
tiktoken>=0.7.0
uvicorn[standard]>=0.30.0